shortened class names
