+++ src/ts/punchcards.ts
@@ -1,10 +1,10 @@
 
 
-export {PunchcardBase} from './punchcards/punchcard-base';
-export {PunchcardColorMap} from './punchcards/punchcard-colormap';
-export {PunchcardDateCircle} from './punchcards/punchcard-date-circle';
-export {PunchcardDateRect} from './punchcards/punchcard-date-rect';
-export {PunchcardLegend} from './punchcards/punchcard-legend';
-export {PunchcardWeekdayCircle} from './punchcards/punchcard-weekday-circle';
-export {PunchcardWeekdayRect} from './punchcards/punchcard-weekday-rect';
+export {Base} from './punchcards/base';
+export {ColorMap} from './punchcards/colormap';
+export {DateCircle} from './punchcards/date-circle';
+export {DateRect} from './punchcards/date-rect';
+export {Legend} from './punchcards/legend';
+export {WeekdayCircle} from './punchcards/weekday-circle';
+export {WeekdayRect} from './punchcards/weekday-rect';
