+++ src/ts/punchcards/legend.ts
@@ -1,17 +1,17 @@
 /// <reference path="../../../typings/globals/d3/index.d.ts" />
 
-import {PunchcardBase} from './punchcard-base';
-import {PunchcardDateCircle} from './punchcard-date-circle';
-import {PunchcardDateRect} from './punchcard-date-rect';
-import {PunchcardWeekdayCircle} from './punchcard-weekday-circle';
-import {PunchcardWeekdayRect} from './punchcard-weekday-rect';
+import {Base} from './base';
+import {DateCircle} from './date-circle';
+import {DateRect} from './date-rect';
+import {WeekdayCircle} from './weekday-circle';
+import {WeekdayRect} from './weekday-rect';
 
 
-type PunchcardVisualization = PunchcardBase|PunchcardDateCircle|PunchcardDateRect|
-                                    PunchcardWeekdayCircle|PunchcardWeekdayRect;
+type PunchcardVisualization = Base|DateCircle|DateRect|
+                                    WeekdayCircle|WeekdayRect;
 
 
-export class PunchcardLegend {
+export class Legend {
 
     private _marginLeft      : number;
     private _marginRight     : number;
@@ -47,7 +47,7 @@ export class PunchcardLegend {
 
 
 
-    public draw(): PunchcardLegend {
+    public draw(): Legend {
         // draw the legend
 
         this.drawLegendBody();
@@ -64,7 +64,7 @@ export class PunchcardLegend {
 
 
 
-    protected drawBox():PunchcardLegend {
+    protected drawBox():Legend {
         // draw box
 
         let dx:number = this.sibling.domElem.clientWidth - this.sibling.legendWidth + this.marginLeft;
@@ -84,7 +84,7 @@ export class PunchcardLegend {
 
 
 
-    private drawHorizontalAxis():PunchcardLegend {
+    private drawHorizontalAxis():Legend {
 
         let dx:number = this.sibling.domElem.clientWidth - this.sibling.legendWidth + this.marginLeft;
         let dy:number = this.sibling.domElem.clientHeight - this.marginBottom;
@@ -105,7 +105,7 @@ export class PunchcardLegend {
 
 
 
-    protected drawLegendBody():PunchcardLegend {
+    protected drawLegendBody():Legend {
         //
         let dx:number = this.sibling.domElem.clientWidth - this.sibling.legendWidth + this.marginLeft;
         let dy:number = this.sibling.domElem.clientHeight - this.marginBottom - this.height;
@@ -125,10 +125,10 @@ export class PunchcardLegend {
 
 
 
-    protected drawSymbols():PunchcardLegend {
+    protected drawSymbols():Legend {
         // pass
 
-        let that:PunchcardLegend = this;
+        let that:Legend = this;
 
         let dx:number = this.sibling.domElem.clientWidth - this.sibling.legendWidth + this.marginLeft;
         let dy:number = this.sibling.domElem.clientHeight - this.marginBottom;
@@ -170,7 +170,7 @@ export class PunchcardLegend {
 
 
 
-    protected drawTitle():PunchcardLegend {
+    protected drawTitle():Legend {
 
         let dx:number = this.sibling.domElem.clientWidth - this.sibling.legendWidth + this.marginLeft + 0.5 * this.width;
         let dy:number = this.marginTop - 20;
@@ -188,7 +188,7 @@ export class PunchcardLegend {
 
 
 
-    protected drawVerticalAxis():PunchcardLegend {
+    protected drawVerticalAxis():Legend {
         //
         let w :number = this.sibling.legendWidth - this.marginLeft - this.marginRight;
         let h :number = this.sibling.domElem.clientHeight - this.marginTop - this.marginBottom;
@@ -217,7 +217,7 @@ export class PunchcardLegend {
 
 
 
-    protected drawVerticalAxisLabel():PunchcardLegend {
+    protected drawVerticalAxisLabel():Legend {
         //
         let h :number = this.sibling.domElem.clientHeight - this.marginTop - this.marginBottom;
         let dx:number = this.sibling.domElem.clientWidth - this.marginRight + 40;
