+++ src/ts/punchcards/weekday-circle.ts
@@ -1,48 +1,45 @@
 /// <reference path="../../../typings/globals/crossfilter/index.d.ts" />
 /// <reference path="../../../typings/globals/d3/index.d.ts" />
-/// <reference path="../../../typings/globals/moment/index.d.ts" />
 
 
-import {PunchcardColorMap} from './punchcard-colormap';
-import {PunchcardDateRect} from './punchcard-date-rect';
+import {ColorMap} from './colormap';
+import {WeekdayRect} from './weekday-rect';
 
 
 
-export class PunchcardDateCircle extends PunchcardDateRect {
-
+export class WeekdayCircle extends WeekdayRect {
 
     constructor (cf: any, domElemId: string) {
 
         super(cf, domElemId);
 
-        this.xlabel = '';
-        this.title = 'PunchcardDateCircle title';
-        this.colormap = new PunchcardColorMap('rainbow');
+        this.xlabel = 'Day of week';
+        this.title = 'WeekdayCircle title';
+        this.colormap = new ColorMap('blues');
     }
 
 
 
 
-    protected drawSymbols():PunchcardDateCircle {
+    protected drawSymbols():WeekdayCircle {
 
         // capture the this object
-        let that:PunchcardDateCircle = this;
+        let that:WeekdayCircle = this;
 
         let w :number = this.domElem.clientWidth - this.marginLeft - this.marginRight - this.legendWidth;
         let h :number = this.domElem.clientHeight - this.marginTop - this.marginBottom;
         let dx:number = this.marginLeft;
         let dy:number = this.marginTop + h;
-        let symbolMargin = {left:0, right: 0, top: 0, bottom: 0}; // pixels
-        let wDays:number = moment(this.dateTo).diff(moment(this.dateFrom), 'days', true);
+        let symbolMargin = {left:2, right: 2, top: 2, bottom: 2}; // pixels
+        let symbolWidth :number = w / 7 - symbolMargin.left - symbolMargin.right;
+        let symbolHeight:number = h / 24 - symbolMargin.top - symbolMargin.bottom;
 
-        let symbolWidth :number = w / wDays - symbolMargin.left - symbolMargin.right;
-        let symbolHeight:number = h / 24.0 - symbolMargin.top - symbolMargin.bottom;
         let r:number = Math.min(symbolWidth, symbolHeight) / 2 - 2;
 
         // based on example from
         // http://stackoverflow.com/questions/16766986/is-it-possible-to-group-by-multiple-dimensions-in-crossfilter
         // forEach method could be very expensive on write.
-        let group = this.dim.dateAndHourOfDay.group();
+        let group = this.dim.weekdayAndHourOfDay.group();
         group.all().forEach(function(d) {
             //parse the json string created above
             d.key = JSON.parse(d.key);
@@ -77,12 +74,12 @@ export class PunchcardDateCircle extends PunchcardDateRect {
                 .append('circle')
                     .attr('class', 'symbol')
                     .attr('cx', function(d){
-                        return that.dateScale(new Date(d.key.datestr));
-                        })
-                    .attr('cy', function(d){
-                        return that.todScale(parseInt(d.key.hourOfDay, 10)) + symbolMargin.top + symbolHeight / 2;
+                        return that.dayOfWeekScale(d.key['weekday']) + symbolMargin.left;
                     })
-                    .attr('r', function(d) {
+                    .attr('cy', function(d){
+                        return that.todScale(d.key['hourOfDay']) + symbolHeight / 2 + symbolMargin.top;
+                    })
+                    .attr('r', function(d){
                         return Math.max(r * (d.value - that.colormap.cLimLow) / (that.colormap.cLimHigh - that.colormap.cLimLow), 1);
                     })
                     .attr('fill', function(d){
@@ -90,8 +87,8 @@ export class PunchcardDateCircle extends PunchcardDateRect {
                     });
 
         return this;
-
     }
+
 
 
 }
